+++ src/EulerHS/Logger/Runtime.hs
@@ -5,6 +5,7 @@ module EulerHS.Logger.Runtime
     -- * Core Runtime
     CoreRuntime(..)
   , LoggerRuntime(..)
+  , SeverityCounterHandle(..)
   , shouldLogRawSql
   , incLogCounter
   , createCoreRuntime
@@ -29,63 +30,92 @@ import qualified System.Logger as Log
 -- TODO: add StaticLoggerRuntimeContext if we'll need more than a single Bool
 data LoggerRuntime
   = LoggerRuntime
-    { _flowFormatter    :: T.FlowFormatter
-    , _logContext       :: T.LogContext
-    , _logLevel         :: T.LogLevel
-    , _logRawSql        :: T.ShouldLogSQL
-    , _logCounter       :: !T.LogCounter
-    , _logMaskingConfig :: Maybe T.LogMaskingConfig
-    , _logLoggerHandle  :: Impl.LoggerHandle
+    { _flowFormatter          :: T.FlowFormatter
+    , _logContext             :: T.LogContext
+    , _logLevel               :: T.LogLevel
+    , _logRawSql              :: T.ShouldLogSQL
+    , _logCounter             :: !T.LogCounter
+    , _logMaskingConfig       :: Maybe T.LogMaskingConfig
+    , _logLoggerHandle        :: Impl.LoggerHandle
+    , _severityCounterHandle  :: Maybe SeverityCounterHandle
     }
-  | MemoryLoggerRuntime !T.FlowFormatter T.LogContext !T.LogLevel !(MVar [Text]) !T.LogCounter
+  | MemoryLoggerRuntime
+      !T.FlowFormatter
+       T.LogContext
+      !T.LogLevel
+      !(MVar [Text])
+      !T.LogCounter
 
 newtype CoreRuntime = CoreRuntime
     { _loggerRuntime :: LoggerRuntime
     }
 
--- createLoggerRuntime :: LoggerConfig -> IO LoggerRuntime
--- createLoggerRuntime (MemoryLoggerConfig cfgLogLevel) =
---   MemoryLoggerRuntime cfgLogLevel <$> newMVar []
--- createLoggerRuntime cfg = do
---   counter <- initLogCounter
---   LoggerRuntime (_level cfg) (_logRawSql cfg) counter Nothing Nothing (_logMaskingConfig cfg)<$> Impl.createLogger cfg
+-- | Log entry counter by severity handle.
+data SeverityCounterHandle = SeverityCounterHandle
+  { incCounter :: T.LogLevel -> IO ()
+  }
 
 createMemoryLoggerRuntime :: T.FlowFormatter -> T.LogLevel -> IO LoggerRuntime
 createMemoryLoggerRuntime flowFormatter logLevel =
   MemoryLoggerRuntime flowFormatter mempty logLevel <$> newMVar [] <*> initLogCounter
 
-createLoggerRuntime :: T.FlowFormatter -> T.LoggerConfig -> IO LoggerRuntime
-createLoggerRuntime flowFormatter cfg = do
-  counter <- initLogCounter
-  LoggerRuntime flowFormatter mempty (T._logLevel cfg) (T._logRawSql cfg) counter Nothing
-    <$> Impl.createLogger flowFormatter cfg
+createLoggerRuntime
+  :: T.FlowFormatter
+  -> Maybe SeverityCounterHandle
+  -> T.LoggerConfig
+  -> IO LoggerRuntime
+createLoggerRuntime flowFormatter severityCounterHandler cfg = do
+  -- log entries' sequential number
+  logSequence <- initLogCounter
+  logHandle <- Impl.createLogger flowFormatter cfg
+  pure $ LoggerRuntime
+    flowFormatter
+    mempty
+    (T._logLevel cfg)
+    (T._logRawSql cfg)
+    logSequence
+    Nothing
+    logHandle
+    severityCounterHandler
 
-createLoggerRuntime'
-  :: Maybe Log.DateFormat
+createLoggerRuntime' :: Maybe Log.DateFormat
   -> Maybe Log.Renderer
   -> T.BufferSize
   -> T.FlowFormatter
+  -> Maybe SeverityCounterHandle
   -> T.LoggerConfig
   -> IO LoggerRuntime
-createLoggerRuntime' mbDateFormat mbRenderer bufferSize flowFormatter cfg = do
-  counter <- initLogCounter
+createLoggerRuntime' mbDateFormat mbRenderer bufferSize flowFormatter severityCounterHandler cfg = do
+  -- log entries' sequential number
+  logSequence <- initLogCounter
   loggerHandle <- Impl.createLogger' mbDateFormat mbRenderer bufferSize flowFormatter cfg
   pure $ LoggerRuntime
     flowFormatter
     mempty
     (T._logLevel cfg)
     (T._logRawSql cfg)
-    counter
+    logSequence
     (T._logMaskingConfig cfg)
     loggerHandle
+    severityCounterHandler
 
 createVoidLoggerRuntime :: IO LoggerRuntime
 createVoidLoggerRuntime = do
-  counter <- initLogCounter
-  LoggerRuntime (const $ pure T.showingMessageFormatter) mempty T.Debug T.SafelyOmitSqlLogs counter Nothing <$> Impl.createVoidLogger
+  -- log entries' sequential number
+  logSequence <- initLogCounter
+  logHandle <- Impl.createVoidLogger
+  pure $ LoggerRuntime
+    (const $ pure T.showingMessageFormatter)
+    mempty
+    T.Debug
+    T.SafelyOmitSqlLogs
+    logSequence
+    Nothing
+    logHandle
+    Nothing
 
 clearLoggerRuntime :: LoggerRuntime -> IO ()
-clearLoggerRuntime (LoggerRuntime flowFormatter _ _ _ _ _ handle) = Impl.disposeLogger flowFormatter handle
+clearLoggerRuntime (LoggerRuntime flowFormatter _ _ _ _ _ handle _) = Impl.disposeLogger flowFormatter handle
 clearLoggerRuntime (MemoryLoggerRuntime _ _ _ msgsVar _) = void $ swapMVar msgsVar []
 
 createCoreRuntime :: LoggerRuntime -> IO CoreRuntime
@@ -96,13 +126,13 @@ clearCoreRuntime _ = pure ()
 
 shouldLogRawSql :: LoggerRuntime -> Bool
 shouldLogRawSql = \case
-  (LoggerRuntime _ _ _ T.UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION _ _ _) -> True
-  _                                                                   -> False
+  (LoggerRuntime _ _ _ T.UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION _ _ _ _) -> True
+  _                                                                     -> False
 
 getLogMaskingConfig :: LoggerRuntime -> Maybe T.LogMaskingConfig
 getLogMaskingConfig = \case
-  (LoggerRuntime _ _ _ _ _ mbMaskConfig _) -> mbMaskConfig
-  _                                        -> Nothing
+  (LoggerRuntime _ _ _ _ _ mbMaskConfig _ _) -> mbMaskConfig
+  _                                          -> Nothing
 
 initLogCounter :: IO T.LogCounter
 initLogCounter = newIORef 0