+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog for euler-hs
 
+## [2.11.0.0] - 2022-02-01
+
+* Add handler to count log entries by severity. It is a BREAKING change! From now on you have to provide a `SeverityCounterHandle` when building logger runtime. Please refer to this PR as an example https://bitbucket.org/juspay/euler-api-customer/pull-requests/249
+
 ## [2.10.0.0] - 2022-01-25
 
 * !!!BREAKING CHANGES!!! Remove non-total functions from re-export.
   Highly recommended to not use them in your code.
-  However in case they are indispensable to code just import from standart Prelude.
+  However, in case they are indispensable to code just import from standart Prelude.
 
 ## [2.9.4.3] - 2022-01-24